+++ src/Parametizer/CliRequest/CliRequest.php
@@ -2,8 +2,8 @@
 
 namespace MagicPush\CliToolkit\Parametizer\CliRequest;
 
+use LogicException;
 use MagicPush\CliToolkit\Parametizer\Config\Config;
-use MagicPush\CliToolkit\Parametizer\Exception\ConfigException;
 use MagicPush\CliToolkit\Parametizer\HelpFormatter;
 
 class CliRequest {
@@ -12,20 +12,28 @@ class CliRequest {
      */
     public function __construct(protected readonly Config $config, protected readonly array $params = []) { }
 
-    public function getParam(string $key): mixed {
-        if (!array_key_exists($key, $this->params)) {
-            throw new ConfigException(
-                "Parameter '" . HelpFormatter::createForStdErr()->paramTitle($key) . "' not found in the request",
+    /**
+     * @return mixed[]
+     */
+    public function getParams(): array {
+        return $this->params;
+    }
+
+    public function getParam(string $paramName): mixed {
+        if (!array_key_exists($paramName, $this->params)) {
+            throw new LogicException(
+                "Parameter '" . HelpFormatter::createForStdErr()->paramTitle($paramName) . "' not found in the request."
+                    . ' The parameters being parsed: ' . implode(', ', array_keys($this->params)),
             );
         }
 
-        return $this->params[$key];
+        return $this->params[$paramName];
     }
 
     public function getCommandRequest(string $command): static {
         $subcommandConfig = $this->config->getBranch($command);
         if (!$subcommandConfig) {
-            throw new ConfigException(
+            throw new LogicException(
                 "Subcommand '" . HelpFormatter::createForStdErr()->paramTitle($command) . "' not found",
             );
         }
@@ -33,10 +41,71 @@ class CliRequest {
         return new static($subcommandConfig, $this->getParam($command));
     }
 
+    private static function validateValueIsArray(string $paramName, mixed $paramValue): void {
+        if (is_array($paramValue)) {
+            return;
+        }
+
+        throw new LogicException(
+            "Parameter '" . HelpFormatter::createForStdErr()->paramTitle($paramName) . "' contains a single value",
+        );
+    }
+
+    private static function validateValueNotArray(string $paramName, mixed $paramValue): void {
+        if (!is_array($paramValue)) {
+            return;
+        }
+
+        throw new LogicException(
+            "Parameter '" . HelpFormatter::createForStdErr()->paramTitle($paramName) . "' contains an array",
+        );
+    }
+
+    public function getParamAsInt(string $paramName): int {
+        $paramValue = $this->getParam($paramName);
+        self::validateValueNotArray($paramName, $paramValue);
+
+        return (int) $paramValue;
+    }
+
     /**
-     * @return mixed[]
+     * @return int[]
      */
-    public function getParams(): array {
-        return $this->params;
+    public function getParamAsIntList(string $paramName): array {
+        $paramValue = $this->getParam($paramName);
+        self::validateValueIsArray($paramName, $paramValue);
+
+        array_walk(
+            $paramValue,
+            function (&$elementValue) {
+                $elementValue = (int) $elementValue;
+            },
+        );
+
+        return $paramValue;
+    }
+
+    public function getParamAsFloat(string $paramName): float {
+        $paramValue = $this->getParam($paramName);
+        self::validateValueNotArray($paramName, $paramValue);
+
+        return (float) $paramValue;
+    }
+
+    /**
+     * @return float[]
+     */
+    public function getParamAsFloatList(string $paramName): array {
+        $paramValue = $this->getParam($paramName);
+        self::validateValueIsArray($paramName, $paramValue);
+
+        array_walk(
+            $paramValue,
+            function (&$elementValue) {
+                $elementValue = (float) $elementValue;
+            },
+        );
+
+        return $paramValue;
     }
 }