+++ tests/Tests/Parametizer/CliRequest/scripts/get-single-as-array.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+
+use MagicPush\CliToolkit\Parametizer\Parametizer;
+
+require_once __DIR__ . '/../../../init-console.php';
+
+$request = Parametizer::newConfig()
+    ->newOption('--option-single')
+
+    ->run();
+
+$request->getParamAsIntList('option-single');