+++ tests/Tests/TerminalFormatter/scripts/no-style-when-not-in-terminal.php
@@ -2,7 +2,7 @@
 
 use MagicPush\CliToolkit\TerminalFormatter;
 
-require_once __DIR__ . '/../init-console.php';
+require_once __DIR__ . '/../../init-console.php';
 
 echo 'some string' . PHP_EOL;
 echo TerminalFormatter::createForStdOut()->apply('some stylish string', [TerminalFormatter::FONT_LIGHT_GREEN])